+++ app/src/main/kotlin/dev/patrickgold/florisboard/lib/compose/FlorisScreen.kt
@@ -223,6 +223,9 @@ private class FlorisScreenScopeImpl : FlorisScreenScope {
                     fab()
                 }
             },
+            snackbarHost = {
+                FlorisSnackbarController.Host()
+            },
         ) { innerPadding ->
             val scrollModifier = if (scrollable) {
                 Modifier.florisVerticalScroll()